2 year/Part 2/Testing/test_lab6/requirements.txt
selenium~=4.38.0
pytest~=8.4.2
allure-pytest~=2.15.0
webdriver-manager~=4.0.2